Fixed issue with config when hypen in Jira state
Fixed issue with config when hypen in Jira state name. This was done by
changing the js toId and fromId methods to use a double underscore as a
placeholder for spaces rather than a hyphen which could also be in a
state's name. Then ran into a similar issue with capitalization: the
toId and fromId methods convert: In Process properly but In-Process is
converted to In-process and it is not recognized as being mapped to a
lane when stepping through the configuration. The fix was to make the
validation case-insensitive.