Updates to github connection to catch auth issues
When connecting to github it only validates the username/password is
valid, it doesn't check the repo is valid. So once you try to retrieve
projects (repos) if the repo is invalid it fails silently. Updated the
code to throw and catch so it will show an error in the UI.